Moved path finding related classes in their own sub folder
