+++ Source/SVONavigation/Public/PathFinding/SVOPathFinder.h
